ioctl: FBTAP_IOCGSIZE type width.
FBTAP_IOCGSIZE writes the number of pages used for the framebuffer in an
unsigned long.


OXT-1638

+++ fbtap.c
@@ -201,7 +201,7 @@ fbtap_ioctl (struct file *filp, unsigned int cmd, unsigned long arg)
         break;
 
     case FBTAP_IOCGSIZE:
-        put_user (fb->size >> PAGE_SHIFT, (unsigned int*) arg);
+        put_user (fb->size >> PAGE_SHIFT, (unsigned long *)arg);
         break;
 
     case FBTAP_IOCALLOCFB:
